Reserv_database_to_dataframe/requirements.txt
mysql-connector-python~=8.0.32
pandas~=1.5.3
SQLAlchemy~=2.0.7